refactor: Move gulp task into it's own file. Create folder
