+++ gulpfile.js
@@ -28,55 +28,11 @@ require('./gulpTasks/copy-assets.js')(gulp, config);
 require('./gulpTasks/release.js')(gulp, creds);
 require('./gulpTasks/compile-html.js')(gulp);
 require('./gulpTasks/local-testing.js')(gulp, config);
+require('./gulpTasks/new-component.js')(gulp, argv);
 
 /* ============================================================ *\
 	MAIN TASKS
 \* ============================================================ */
-
-gulp.task('component', function(cb) {
-
-	var fs = require('fs');
-	var path = require('path');
-	
-	var viewsPath = path.join(process.cwd(), 'views')
-	var partialsPath = path.join(viewsPath, '_partials');
-	var componentNewPath = path.join(partialsPath, argv.name)
-
-	var hbsFileName;
-	var scssFileName;
-
-	if(!argv.name) {
-		console.log('');
-		console.log('A component must have a name');
-		console.log('');
-		console.log('Example usage:')
-		console.log('gulp component --name header');
-		console.log('');
-		cb();
-	} else {
-
-		console.log('time to create a component name: %s', argv.name);
-
-		hbsFileName = argv.name + '.hbs';
-		scssFileName = argv.name + '.scss';
-
-		// fs.mkdir(viewsPath, function(err) {
-
-		// })
-
-		console.log('create %s folder', componentNewPath);
-		console.log('create %s', hbsFileName);
-		console.log('create %s', scssFileName);
-
-		//@TODO 
-		//create folder
-		//create handlebars file
-		//create sass partial
-
-		cb();
-	}
-
-})
 
 gulp.task('watch:sass', function () {
 	if(!argv.prod) {
@@ -104,6 +60,10 @@ gulp.task('watch:sprites', function () {
 		);
 	}
 });
+
+gulp.task('component', function(cb) {
+	runSeq(['new-component'], cb);
+})
 
 gulp.task('watch', function (cb) {
 	runSeq(['watch:sass', 'watch:js', 'watch:sprites'], cb);
